% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{antibodyData}
\alias{antibodyData}
\title{Example antibody data}
\format{A data frame with 256 rows and 14 cols where the first column gives the patient ID (blanks are filled down), the second column gives the antigen and columns 3-14 give OD values from 96 well plates.}
\source{
system.file("data-raw", "makeExample.R", package = "antir")
}
\usage{
antibodyData
}
\description{
A dataset containing raw data read in from the .csv results from an example antibody experiment
}
\keyword{datasets}
