% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotAnti}
\alias{plotAnti}
\title{Plot antibody profiling data}
\usage{
plotAnti(stacked, positiveThreshold = 25,
  lab = "Dilution(OD450=p24 25pg)")
}
\arguments{
\item{stacked}{output from readAnti storing antibody profiling data}

\item{positiveThreshold}{a single numeric threshold to use for determining a threshold by estimating the OD for this amount of positive control antibody}

\item{lab}{a string giving the label for the color scale}
}
\description{
Plot antibody profiling data from 96 well plates using positive and negative controls and replicate serial dilutions of patient plasma.
}
\examples{
anti<-readAnti(antibodyData)
plotAnti(anti[anti$plate<4,])
}
\seealso{
\code{\link{plotAnti}}
}
