% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{readAnti}
\alias{readAnti}
\title{Read antibody profiling data}
\usage{
readAnti(dat, positiveThreshold = 25, dilutions = 300 * 3^(0:3),
  p24Dilutions = 240/3^(0:3), vocal = TRUE, nrows = 8)
}
\arguments{
\item{dat}{a file containing the raw data or a data.frame containing raw data. The function assumes that the first column of the first row of each plate contains the patient ID and the remainder of that plate is empty.}

\item{positiveThreshold}{a single numeric threshold to use for determining a threshold by estimating the OD for this amount of positive control antibody}
}
\value{
a data.frame
}
\description{
Read in antibody profiling data from 96 well plates using positive and negative controls and replicate serial dilutions of patient plasma
}
\examples{
fakeDat<-1:8 \%*\% t(1/3^rep(0:3,3))
fakeDat[1,]<-c(rep(0.1,4),50*1/3^rep(0:3,2))
raw<-cbind(data.frame('pat'='Test',antigen=LETTERS[1:8]),fakeDat)
tmpFile<-tempfile()
write.table(raw,tmpFile,row.names=FALSE,col.names=FALSE,sep=',')
readAnti(tmpFile)
readAnti(raw)
}
\seealso{
\code{\link{calcP24Cut}}, \code{\link{calcCross}}, \code{\link{plotAnti}}
}
